+++ src/frontend/src/blocks/blocksHandler/triggers/triggerEvents.ts
@@ -3,13 +3,11 @@ import { discordExportEvents, discordInputFields, discordTriggerEvents,  } from 
 import { githubExportEvents, githubInputFields, githubTriggerEvents } from "../../github/triggerConfig";
 import { rssExportEvents, rssInputFields, rssTriggerEvents } from "../../rss/triggerConfig";
 import { telegramExportEvents, telegramInputFields, telegramTriggerEvents, } from "../../telegram/triggerConfig";
-import { walletExportEvents, walletInputFields, walletTriggerEvents } from "../../web3 wallet/triggerConfig";
 
 export const triggerEventsByApp: Record<string, Array<{ value: string; label: string; icon: any }>> = {
   telegram: telegramTriggerEvents,
   discord: discordTriggerEvents,
   github: githubTriggerEvents,
-  wallet: walletTriggerEvents,
   rss: rssTriggerEvents,
 };
 
@@ -17,7 +15,6 @@ export const exportEventsByApp: Record<string, Array<{ value: string; label: str
   ...telegramExportEvents,
   ...discordExportEvents,
   ...githubExportEvents,
-  ...walletExportEvents,
   ...rssExportEvents,
 };
 
@@ -25,6 +22,5 @@ export const inputFieldsByApp: Record<string, InputField[]> = {
   telegram: telegramInputFields.telegram,
   discord: discordInputFields.discord,
   github: githubInputFields.github,
-  wallet: walletInputFields.wallet,
   rss: rssInputFields.rss,
 };
