+++ src/frontend/src/components/workflowBuilder/blocks.ts
@@ -6,7 +6,7 @@ import email from "../../assets/email.png"
 // import metamask from "../../assets/metamask.png"
 // import notion from "../../assets/notion.png"
 import telegram from "../../assets/telegram.png"
-import etherium from "../../assets/eth.png"
+import token from "../../assets/token.png"
 //import twitter from "../../assets/twitter.jpg"
 import ai from "../../assets/ai.png"
 import type { AppBlock } from "./types"
@@ -18,7 +18,7 @@ export const appBlocks: AppBlock[] = [
   { type: "email", label: "Email", icon: Mail, iconUrl: email, color: "white", category: "action" },
   { type: "github", label: "GitHub", icon: Github, iconUrl: github, color: "white", category: "trigger" },
   // { type: "wallet", label: "Wallet Tracking", icon: TrendingUp, iconUrl: metamask, color: "white", category: "trigger"},
-  { type: "web3Token", label: "Token Price", icon: Coins , iconUrl: etherium, color: "white", category: "action" },
+  { type: "web3Token", label: "Token Price", icon: Coins , iconUrl: token, color: "white", category: "action" },
   // { type: "rss", label: "RSS Feed", icon: Coins , iconUrl: etherium, color: "white", category: "trigger" },
   { type: "discord",label: "Discord",icon: MessageSquare,iconUrl: discord,color: "white",category: "action"},
   // { type: "notion",label: "Notion",icon: FileText,iconUrl: notion,color: "white",category: "action"},
